+++ src/main/java/com/gregtechceu/gtceu/data/lang/ArmorLang.java
@@ -15,6 +15,7 @@ public class ArmorLang {
         provider.add("item.gtceu.armor.boots", "%s Boots");
     }
 
+
     private static void generateArmorMessageKeys(GTLangProvider provider) {
         // NanoMuscle Suit
         provider.add("armor.gtceu.nano_muscle_suite.night_vision.enabled", "NanoMuscle™ Suite: §aNightVision Enabled");
