+++ src/main/java/com/gregtechceu/gtceu/data/lang/LangHandler.java
@@ -1,11 +1,13 @@
 package com.gregtechceu.gtceu.data.lang;
 
 import com.gregtechceu.gtceu.api.registry.registrate.provider.GTLangProvider;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.Style;
 
 public class LangHandler {
 
     /**
-     *
+     * MC language formatting is moving towards using {@link MutableComponent#withStyle(Style)} instead of § codes, so {@link MutableComponent#withStyle(Style)} should be used where possible.<br>
      * Unformatted text must be in light gray - §7
      * Items must be in gold - §6
      * Fluids must be in blue - §9
@@ -17,10 +19,11 @@ public class LangHandler {
      * Percentages must be in green - §a
      * Keys must be in all caps
      * Key combos must follow the format KEY1 + KEY2 (for example, SHIFT + R-CLICK)
-     *
      */
 
     public static void init(GTLangProvider provider) {
+
+        CommonLang.init(provider);
         AdvancementLang.init(provider);
         ArmorLang.init(provider);
         BlockLang.init(provider);
