+++ src/main/java/com/gregtechceu/gtceu/data/lang/MaterialLang.java
@@ -1,19 +1,22 @@
 package com.gregtechceu.gtceu.data.lang;
 
 import com.gregtechceu.gtceu.api.data.chemical.material.Material;
-import com.gregtechceu.gtceu.api.data.chemical.material.registry.MaterialRegistry;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
 import com.gregtechceu.gtceu.api.registry.registrate.provider.GTLangProvider;
 
 import java.util.stream.Collectors;
 
-import static com.gregtechceu.gtceu.common.data.GTMaterials.*;
-
 public class MaterialLang {
 
-    public static void generate(GTLangProvider provider, String namespace) {
+    public static void generateMaterialLang(GTLangProvider provider, String namespace) {
         for (Material material : GTRegistries.MATERIALS.values().stream().filter(v -> v.getModid().equals(namespace)).collect(Collectors.toSet())) {
             provider.add(material.getUnlocalizedName(), material.getDefaultTranslation());
+
+            for (var entry: material.getLangOverrides().entrySet()) {
+                var key = String.format("item.%s.%s", material.getResourceLocation().getNamespace(),
+                        entry.getKey().idPattern().formatted(material.getResourceLocation().getPath()));
+                provider.add(key, entry.getValue());
+            }
         }
     }
