+++ src/main/java/com/gregtechceu/gtceu/data/lang/MultiblockLang.java
@@ -1,10 +1,11 @@
 package com.gregtechceu.gtceu.data.lang;
 
+import com.gregtechceu.gtceu.api.registry.registrate.provider.GTLangProvider;
 import com.tterrag.registrate.providers.RegistrateLangProvider;
 
 public class MultiblockLang {
 
-    public static void init(RegistrateLangProvider provider) {
+    public static void init(GTLangProvider provider) {
         provider.add("gtceu.pattern_error.mismatch_coils", "Mismatched coils: %s vs %s at (%d, %d, %d)");
         provider.add("gtceu.pattern_error.mismatch_filters", "Mismatched filters: %s vs %s at (%d, %d, %d)");
         provider.add("gtceu.predicate_error.active_transformer.missing_io", "Missing hatches: IN - %s | OUT - %s");
@@ -23,5 +24,23 @@ public class MultiblockLang {
         provider.add("gtceu.predicate_error.charcoal.logs", "Must be completely filled with logs.");
         provider.add("gtceu.predicate_error.steam.missing_steam_hatch", "Missing steam hatch");
         provider.add("gtceu.pattern_predicate.blocks", "Error at X: %d, Y: %d, Z: %d");
+
+        // Multiblock Preview
+        provider.add("gtceu.multiblock.preview.zoom", "Use mousewheel or right-click + drag to zoom");
+        provider.add("gtceu.multiblock.preview.rotate", "Click and drag to rotate");
+        provider.add("gtceu.multiblock.preview.select", "Right-click to check candidates");
+        provider.add("gtceu.multiblock.pattern.error", "Expected components (%s) at (%s).");
+        provider.add("gtceu.multiblock.pattern.error.limited_exact", "§cExactly: %d§r");
+        provider.add("gtceu.multiblock.pattern.error.limited_within", "§cBetween %d and %d§r");
+        provider.addMultiLang("gtceu.multiblock.pattern.error.limited", "§cMaximum: %d§r", "§cMinimum: %d§r",
+                "§cMaximum: %d per layer§r", "§cMinimum: %d per layer§r");
+        provider.add("gtceu.multiblock.pattern.error.coils", "§cAll heating coils must be the same§r");
+        provider.add("gtceu.multiblock.pattern.error.filters", "§cAll filters must be the same§r");
+        provider.add("gtceu.multiblock.pattern.error.batteries", "§cAll batteries must be the same§r");
+        provider.add("gtceu.multiblock.pattern.clear_amount_1", "§6Must have a clear 1x1x1 space in front§r");
+        provider.add("gtceu.multiblock.pattern.clear_amount_3", "§6Must have a clear 3x3x1 space in front§r");
+        provider.add("gtceu.multiblock.pattern.single", "§6Only this block can be used§r");
+        provider.add("gtceu.multiblock.pattern.location_end", "§cVery End§r");
+        provider.add("gtceu.multiblock.pattern.replaceable_air", "Replaceable by Air");
     }
 }
