+++ src/main/java/com/gregtechceu/gtceu/integration/jade/provider/ControllableBlockProvider.java
@@ -41,9 +41,9 @@ public class ControllableBlockProvider extends CapabilityBlockProvider<IControll
                               BlockEntity blockEntity, IPluginConfig config) {
         if (capData.contains("SuspendAfter") && capData.getBoolean("SuspendAfter")) {
             tooltip.add(
-                    Component.translatable("item.gtceu.behavior.soft_hammer.disabled_cycle").withStyle(ChatFormatting.YELLOW));
+                    Component.translatable("common.gtceu.workable.disabled_next_cycle").withStyle(ChatFormatting.YELLOW));
         } else if (capData.contains("WorkingEnabled") && !capData.getBoolean("WorkingEnabled")) {
-            tooltip.add(Component.translatable("item.gtceu.behavior.soft_hammer.disabled").withStyle(ChatFormatting.YELLOW));
+            tooltip.add(Component.translatable("common.gtceu.workable.disabled").withStyle(ChatFormatting.YELLOW));
         }
     }
 }
